Реализация тестирования декоратора log() и внесение обновление файла README с информацией о модуле decorators
