+++ README.md
@@ -14,7 +14,8 @@
 - *card_number_generator* - генерирует номера карт 
 в заданном диапазоне от 1 до 9999999999999999 в заданном формате XXXX XXXX XXXX XXXX
 
-
+В модуле **decorators.py** реализован декоратор *log()*, который автоматически регистрирует детали выполнения
+функций, а именно имя функции, время и результат ее выполнения, информацию об ошибках. 
 ## Установка ##
 1. Клонируйте проект в свой репозиторий по ссылке:
 ```
