+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.12">coverage.py v7.6.12</a>,
-            created at 2025-02-16 15:32 +0300
+            created at 2025-03-02 00:27 +0300
         </p>
     </div>
 </header>
@@ -76,6 +76,20 @@
                 <td>0</td>
                 <td>0</td>
                 <td class="right" data-ratio="0 0">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_decorators_py.html">src\decorators.py</a></td>
+                <td>19</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="19 19">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_generators_py.html">src\generators.py</a></td>
+                <td>21</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="21 21">100%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_masks_py.html">src\masks.py</a></td>
@@ -102,10 +116,10 @@
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>51</td>
+                <td>91</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="51 51">100%</td>
+                <td class="right" data-ratio="91 91">100%</td>
             </tr>
         </tfoot>
     </table>
@@ -117,7 +131,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.12">coverage.py v7.6.12</a>,
-            created at 2025-02-16 15:32 +0300
+            created at 2025-03-02 00:27 +0300
         </p>
     </div>
     <aside class="hidden">
