+++ log.txt
@@ -1,0 +1,6 @@
+my_multiplication ok Lead time: 0:00:00.000002
+test_func2 ok Lead time: 0:00:00.000002
+test_func2 ok Lead time: 0:00:00.000001
+test_func2 ok Lead time: 0:00:00.000001
+test_func2 ok Lead time: 0:00:00.000002
+test_func2 ok Lead time: 0:00:00.000001
